refactor(task-engine): drop the task-runner back-compat barrel and shims
Clean break — no backward compatibility:
- Delete server/task-runner.ts (was a pure re-export barrel) and repoint all
  30 importers + test mocks to the real entry server/task-engine/index.js.
- Remove the verify.ts incrementConversationUsage/getConversationUsage delegating
  wrappers; callers + checkUsageGuardrail use the store helpers directly; drop the
  now-duplicated usage tests from verify.test.ts (covered by store.test.ts).
- Retire the stale childLogger('task-runner') labels in lifecycle.ts/cleanup.ts
  in favor of task-engine/<module>.

Behavior-neutral; full suite green (3302).
