+++ server/api.soft-delete.test.ts
@@ -6,7 +6,7 @@ import { createApp } from './app.js';
 
 // ─── Mocks ────────────────────────────────────────────────────────────────────
 
-vi.mock('./task-runner.js', async () => {
+vi.mock('./task-engine/index.js', async () => {
   const { getDb } = await import('./db.js');
   return {
     startTask: vi.fn(),
