+++ server/services/task-service.test.ts
@@ -15,7 +15,7 @@ import { getDb } from '../db.js';
 // ─── Mock side-effecting deps (mirror orchestrator/exec.test.ts) ────────────
 
 const mockStartTask = vi.fn().mockResolvedValue(undefined);
-vi.mock('../task-runner.js', () => ({
+vi.mock('../task-engine/index.js', () => ({
   startTask: vi.fn((...args: unknown[]) => mockStartTask(...args)),
 }));
