+++ server/services/task-service.ts
@@ -13,7 +13,7 @@ import {
   inTransaction,
 } from '../repositories/index.js';
 import { upsertManagedTask } from '../orchestrator/store.js';
-import { startTask } from '../task-runner.js';
+import { startTask } from '../task-engine/index.js';
 import { broadcast } from '../events.js';
 import { childLogger } from '../logger.js';
 import { ServiceError } from './errors.js';
