--- server/task-runner.ts
@@ -1,6 +1,0 @@
-// Back-compat barrel — all symbols live in server/task-engine/.
-// Every external caller (routes, poller, teams, orchestrator, index.ts)
-// continues to resolve its imports unchanged.
-export * from './task-engine/index.js';
-export type { AddAgentOpts } from './task-engine/index.js';
-export type { UserTerminalResult } from './task-engine/index.js';
